+++ arviz/tests/backend_tests/test_data_pyro.py
@@ -7,7 +7,7 @@ import pyro
 from pyro.infer import Predictive
 
 from ..data.io_pyro import from_pyro
-from .helpers import (  # pylint: disable=unused-import
+from ..helpers import (  # pylint: disable=unused-import
     chains,
     check_multiple_attrs,
     draws,
