+++ arviz/tests/base_tests/test_plot_utils.py
