+++ arviz/tests/base_tests/test_plots_bokeh.py
@@ -7,7 +7,7 @@ import numpy as np
 import pytest
 
 from ..data import from_dict, load_arviz_data
-from .helpers import (  # pylint: disable=unused-import
+from ..helpers import (  # pylint: disable=unused-import
     eight_schools_params,
     models,
     create_model,
