+++ arviz/tests/base_tests/test_utils.py
